Prefix all function calls with a \
This means PHP won't search the current namespace for a function first
It also means PHP can use the Opcache (opcode?) for some functions,
which is the actual speed increase.

If possible i'd use the `use function` syntax, but that became
available in php 5.6, and thus isn't an option atm.