Add more psalm assertions (#143)
* Add annotation to valid array key

This allows psalm to understand the method validates the input
as correct array keys

* Ass psalm annotation to is list

This allows psalm to understand the output is a list.
This does mean we bump the minimum version of psalm we are compatible
with to 3.6. Otherwise it could interpet this as thinking the
isList makes the value is the `list` class.