Removing `@psalm-assert string|object` from `ip`, `ipv4`, `ipv6` and `email` validators (for now)
Psalm currently *cannot* deal with union types in `@psalm-assert`, because all `@psalm-assert` expressions
are implicitly inlined by the tool as `assert(type)` language expression.

This means that `@psalm-assert string|object $foo` becomes `assert(is_string($foo) || is_object($foo))`,
which is non-sensical in contexts where `$foo` may never be a `string` or an `object`, leading to impossible
to avoid `TypeDoesNotContainType` or `RedundantCondition` reported psalm errors.

Therefore, dropping the type assertion is a simplistic (for now) solution, hoping that upstream tooling will
someday be able to deal with this sort of assertions in a more atomic manner.

Ref: https://github.com/vimeo/psalm/issues/2452#issuecomment-564059160
Ref: https://github.com/vimeo/psalm/issues/2452#issuecomment-564441827