Reduced dependency graph by removing `symfony/polyfill-ctype`, requiring `ext-ctype`
The resaon this change is necessary is that `symfony/polyfill-ctype` is a **provider** of `ext-ctype`,
but it is not `ext-ctype`.

`ext-ctype` is the actual dependency being requested, and it is up to downstream consumers to decide
whether to install the extension, or polyfill it.

With this change, we effectively revert `1.4.0`, and slim down the dependency graph by a bit for
hundreds of thousands of downstream projects.

Ref: https://github.com/webmozarts/assert/commit/18879ed87119d4d11d57eb76c2997200dd632ce1
Ref: https://github.com/webmozarts/assert/commit/18879ed87119d4d11d57eb76c2997200dd632ce1#commitcomment-70510654